test(hwpx): 표·그림 매퍼 공유 픽스처를 분리한다
HwpxObjectMapperTests가 SwiftLint type_body_length error 임계(400줄)에
다가서 새 테스트를 넣을 수 없었다. id 테이블·파서 래퍼·실물 구조
XML(tableXML·pictureXML)을 HwpxObjectFixture로 옮긴다 (헤더·구역
스위트의 같은 전례). 테스트 단언은 그대로다.
